Multiple Changes - See description
Made changes to reduce number of conditional checks being made and streamlined some actions and checks. Reorganized the commenting and section structure of the code. Implemented free movement and edited player to account for each commander, reducing the number of actions if a commander is captured. Grouping pieces with commanders and sharing a command action for pieces in a group. Implemented free commander movement.