Minor Changes (see description)
Added EditorUserSettings.asset, which is a useless file in that always throws conflicts, to the gitignore so that it no longer gets included. I also made it so that the Normal and Hard difficulty buttons at least start the game, so that it is not just the Easy button that starts the game. Actual difficulty will be added later on in development.